requirements.txt
discord>=2.3.2
python-dotenv>=0.21.1
SQLAlchemy>=2.0.18
PyMySQL>=1.0.2
pytz>=2022.6
sqlalchemy_utils
psycopg2
postgres
discord-py-utilities>=0.0.9.7
fastapi